error handler

+++ backend/job/views.py
@@ -1,20 +1,23 @@
 from django.shortcuts import get_object_or_404
-from job.filters import JobFilter
-from job.models import Job
-from job.serializers import JobSerializer
 from rest_framework import status
 from rest_framework.decorators import api_view
+from rest_framework.decorators import permission_classes
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+
+from .filters import JobFilter
+from .models import Job
+from .serializers import JobSerializer
 
 
 @api_view(['GET'])
-def getAllJobs(request):
-    filterset = JobFilter(request.GET, queryset=Job.objects.all().order_by('id'))
-    count = filterset.qs.count()
+def get_all_jobs(request):
+    filter = JobFilter(request.GET, queryset=Job.objects.all().order_by('id'))
+    count = filter.qs.count()
     paginator = PageNumberPagination()
     paginator.page_size = 2
-    queryset = paginator.paginate_queryset(filterset.qs, request)
+    queryset = paginator.paginate_queryset(filter.qs, request)
     serializer = JobSerializer(queryset, many=True)
     return Response({
         'jobs': serializer.data,
@@ -24,7 +27,7 @@ def getAllJobs(request):
 
 
 @api_view(['GET'])
-def getOneJob(request, pk):
+def get_one_job(request, pk):
     job = get_object_or_404(Job, id=pk)
     serializer = JobSerializer(job, many=False)
 
@@ -32,7 +35,9 @@ def getOneJob(request, pk):
 
 
 @api_view(['POST'])
-def createJob(request):
+@permission_classes([IsAuthenticated])
+def create_job(request):
+    request.data['user'] = request.user
     data = request.data
     job = Job.objects.create(**data)
     serializer = JobSerializer(job, many=False)
@@ -41,8 +46,17 @@ def createJob(request):
 
 
 @api_view(['PUT'])
-def updateJob(request, pk):
+@permission_classes([IsAuthenticated])
+def update_job(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response(
+            {
+                'message': 'You can not update this job!'
+            },
+            status=status.HTTP_403_FORBIDDEN
+        )
     job.description = request.data['description']
     job.email = request.data['email']
     job.address = request.data['address']
@@ -64,6 +78,13 @@ def updateJob(request, pk):
 @api_view(['DELETE'])
 def delete_job(request, pk):
     job = get_object_or_404(Job, id=pk)
+    if job.user != request.user:
+        return Response(
+            {
+                'message': 'You can not update this job!'
+            },
+            status=status.HTTP_403_FORBIDDEN
+        )
     job.delete()
 
     return Response({'message': 'Job is deleted.'}, status=status.HTTP_200_OK)
